requirements.txt
streamlit
pandas
PyPDF2
python-docx
gtts
google-generativeai
dotenv
